requirement.txt
IMDbPY
openpyxl
pandas
mysql-connector-python